+++ tlseparation/utility/filtering.py
@@ -155,9 +155,12 @@ def continuity_filter(wood, leaf, rad=0.05):
     # Stacking wood and leaf arrays.
     arr = np.vstack((wood, leaf))
 
+    # Getting root index (base_id) from point cloud 'arr'.
+    base_id = np.argmin(arr[:, 2])
+
     # Calculating shortest path graph over sampled array.
-    G = array_to_graph(arr, 0, 3, 100, 0.05, 0.02, 0.5)
-    node_ids, dist = extract_path_info(G, 0, return_path=False)
+    G = array_to_graph(arr, base_id, 3, 100, 0.05, 0.02, 0.5)
+    node_ids, dist = extract_path_info(G, base_id, return_path=False)
     node_ids = np.array(node_ids)
 
     # Obtaining wood point cloud indices.